getting output, height's are 'reversed' -- meaning root.height == leaf.height but it should be root.height = max height and leaf = 0.

+++ AVL.h
@@ -95,7 +95,7 @@ class AVLnode {
   }
 
  private:
-  long int key_;
+  int key_;
   std::weak_ptr<AVLnode> parent_;
   std::shared_ptr<AVLnode> left_;
   std::shared_ptr<AVLnode> right_;
